+++ src/rpad/pybullet_envs/pm_suction.py
@@ -7,11 +7,11 @@ import numpy as np
 import numpy.typing as npt
 import pybullet as p
 import rpad.visualize_3d.plots as vpl
-from rpad.partnet_mobility_utils.articulate import articulate_joint
 from rpad.partnet_mobility_utils.data import PMObject
 from rpad.partnet_mobility_utils.render.pybullet import PMRenderEnv, PybulletRenderer
 from scipy.spatial.transform import Rotation as R
 
+from rpad.pybullet_envs.flowbot_utils import compute_normalized_flow
 from rpad.pybullet_envs.suction_gripper import FloatingSuctionGripper
 from rpad.pybullet_envs.suction_gripper_v2 import FloatingSuctionGripperV2
 
@@ -265,55 +265,6 @@ def run_trial(
     )
 
 
-# COPIED FROM FLOWBOT, DON'T WANT TO ADD FLOWBOT AS A DEPENDENCY...
-def compute_normalized_flow(
-    P_world, T_world_base, current_jas, pc_seg, labelmap, pm_raw_data, linknames
-):
-    """Compute normalized flow for an object, based on its kinematics.
-
-    Args:
-        P_world (npt.NDArray[np.float32]): Point cloud render of the object in the world frame.
-        T_world_base (npt.NDArray[np.float32]): The pose of the base link in the world frame.
-        current_jas (Dict[str, float]): The current joint angles (easy to acquire from the render that created the points.)
-        pc_seg (npt.NDArray[np.uint8]): The segmentation labels of each point.
-        labelmap (Dict[str, int]): Map from the link name to segmentation name.
-        pm_raw_data (PMObject): The object description, essentially providing the kinematic structure of the object.
-        linknames (Union[Literal['all'], Sequence[str]], optional): The names of the links for which to
-            compute flow. Defaults to "all", which will articulate all of them.
-
-    Returns:
-        npt.NDArray[np.float32]: _description_
-    """
-
-    # We actuate all links.
-    if linknames == "all":
-        joints = pm_raw_data.semantics.by_type("slider")
-        joints += pm_raw_data.semantics.by_type("hinge")
-        linknames = [joint.name for joint in joints]
-
-    flow = np.zeros_like(P_world)
-
-    for linkname in linknames:
-        P_world_new = articulate_joint(
-            pm_raw_data,
-            current_jas,
-            linkname,
-            0.01,  # Articulate by only a little bit.
-            P_world,
-            pc_seg,
-            labelmap,
-            T_world_base,
-        )
-        link_flow = P_world_new - P_world
-        flow += link_flow
-
-    largest_mag: float = np.linalg.norm(flow, axis=-1).max()
-
-    normalized_flow = flow / (largest_mag + 1e-6)
-
-    return normalized_flow
-
-
 # max simulation steps for control
 TIMEOUT = 1000
 
@@ -386,6 +337,7 @@ class PMSuctionDemoEnv:
             for _ in range(n_steps):
                 # control, sim
                 self.gripper.set_pull_cmds(ctrl)
+                # self.gripper.debug_joint_forces()
                 p.stepSimulation(self.client_id)
                 if self.gui:
                     time.sleep(1 / 240.0)
